Version 1.2.0
* Added more further-reading links to `memset_explicit` and `explicit_bzero` (in `list.c`)
* Fixed small memory leaks when failing to open the password file
* Made `bytes_from_heap_data` the only constructor of `Bytes` objects
* Fixed a bug in `bytes_concat` where we would copy the `first`'s data instead of the `second`'s
* Verified that the salting+sha512 function works by testing against python.

+++ src/vtlock.c
@@ -448,7 +448,7 @@ ExitCode actual_main(Args* args) {
             printf("\x1b[K\x1b[0K\x1b[2K");
             printf(CYAN"Press [Enter] to unlock. "RESET);
             fflush(stdout);
-            sleep_milli(1000/args->clock_fps);
+            sleep_milli((unsigned int) 1000/args->clock_fps);
         }
 
         cursor_move(1, 1);
